chore: Extract convert helper
